requirements.txt
asgiref==3.4.1
certifi==2022.6.15
charset-normalizer==2.1.0
coreapi==2.3.3
coreschema==0.0.4
Django==3.2.10
django-cors-headers==3.10.1
django-filter==22.1
django-rest-swagger==2.2.0
django-utils-six==2.0
djangorestframework==3.13.1
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.2.0
idna==3.3
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
openapi-codec==1.3.2
Pillow==9.2.0
psycopg2-binary==2.9.2
PyJWT==1.7.1
pytz==2021.3
requests==2.28.1
simplejson==3.17.6
sqlparse==0.4.2
typing-extensions==4.0.1
uritemplate==4.1.1
urllib3==1.26.10
